ajuste de formatação no gráfico com base nos mockups
