+++ src/screens/ReportsScreen.tsx
@@ -1,20 +1,10 @@
 import React from "react";
-import { BarChart } from "react-native-chart-kit";
-import { SafeAreaView, View, Text, Dimensions, StyleSheet } from "react-native";
-import { ButtonNavigation } from "../components";
+import { SafeAreaView, View, Text, StyleSheet } from "react-native";
+import { ButtonNavigation, Report } from "../components";
 import colors from "../utils/colors";
 import sizes from "../utils/sizes";
 
 const ReportsScreen: React.FC = () => {
-  const data = {
-    labels: ["Menor", "Mediana", "Maior"],
-    datasets: [
-      {
-        data: [10, 15, 26],
-      },
-    ],
-  };
-
   return (
     <SafeAreaView>
       <View>
@@ -27,37 +17,7 @@ const ReportsScreen: React.FC = () => {
         <View>
           <Text style={styles.text}>Atividades mais frequentes:</Text>
         </View>
-        <View>
-          <BarChart
-            data={data}
-            width={
-              Dimensions.get("window").width -
-              Dimensions.get("window").width / 100
-            }
-            height={
-              Dimensions.get("window").height -
-              Dimensions.get("window").height / 2
-            }
-            yAxisLabel=""
-            yAxisSuffix=""
-            chartConfig={{
-              decimalPlaces: 2, // optional, defaults to 2dp
-              color: (opacity = 1) => `rgba(255, 255, 255, ${opacity})`,
-              labelColor: (opacity = 1) => `rgba(255, 255, 255, ${opacity})`,
-              style: {
-                borderRadius: 16,
-                alignSelf: "center",
-                paddingTop: 40,
-              },
-              propsForDots: {
-                r: "6",
-                strokeWidth: "2",
-                stroke: "#ffa726",
-              },
-            }}
-            showBarTops
-          />
-        </View>
+        <Report title="Medir pressão" />
         <View style={styles.base}>
           <ButtonNavigation
             title="Próximo"
